Java/urban-environment/ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
scikit-learn==1.3.2
pandas==2.1.4
psycopg2-binary==2.9.9
joblib==1.3.2
python-dotenv==1.0.0
numpy==1.24.3
pydantic==2.5.0
requests==2.31.0